Implementation : Middleware and Session

+++ routes/web.php
@@ -11,18 +11,26 @@
 |
 */
 
+use App\Http\Middleware\AdminLoginMiddleware;
+
 Route::get('/admin', 'Core\LoginController@showLayout');
 
 Route::post('/admin/login', 'Core\LoginController@login');
 
-Route::get('/admin/provider', 'Core\ProviderController@providershow');
+Route::get('/admin/logout', 'Core\LoginController@logout');
 
-Route::get('/admin/provider/new', 'Core\ProviderController@showNew');
+Route::get('admin/dashboard', function () {
+    return view('dashboard');
+});
 
-Route::post('/admin/createnewprovider', 'Core\ProviderController@provideradd');
+Route::get('/admin/provider', 'Core\ProviderController@providershow')->middleware(AdminLoginMiddleware::class);
 
-Route::delete('/admin/provider/{provider}', 'Core\ProviderController@providerdelete');
+Route::get('/admin/provider/new', 'Core\ProviderController@showNew')->middleware(AdminLoginMiddleware::class);
 
-Route::get('/admin/provider/edit/{id}', 'Core\ProviderController@providerShowOne');
+Route::post('/admin/createnewprovider', 'Core\ProviderController@provideradd')->middleware(AdminLoginMiddleware::class);
 
-Route::post('admin/provider/editprovider/{id}', 'Core\ProviderController@providerEdit');
+Route::delete('/admin/provider/{provider}', 'Core\ProviderController@providerdelete')->middleware(AdminLoginMiddleware::class);
+
+Route::get('/admin/provider/edit/{id}', 'Core\ProviderController@providerShowOne')->middleware(AdminLoginMiddleware::class);
+
+Route::post('admin/provider/editprovider/{id}', 'Core\ProviderController@providerEdit')->middleware(AdminLoginMiddleware::class);
